refactor: npm preset lookup direct
